1.PVE难度曲线微调，末影龙移除坚韧附魔，龙息伤害增强，但是不再能触发附魔. 2.现在每局首次投喂电池只增加指定附魔次数而不再增加额外附魔栏位了。 3.PVE模式中怪物附魔栏位增加上限，最多只能拥有7个附魔栏位。 4.海洋地图增加一起堵门点，大家一起堵门QAQ 5.修复战胜末影龙后游戏不能正常结束的bug 6.血条刻度增加绿色刻度线，一刻度表示2500血。 7.增加交易命令，输入交易圣诞树可中下圣诞树，一颗圣诞树需要180绿宝石，输入交易列表可查看可交易的道具。 8.更新说明Ui 9.凋零削弱，凋零头颅照成的爆炸伤害由攻击力*2调整为攻击力*1.5 9.召唤法杖重做，每5秒有20%概率获得1个鬼火，之后发送召唤僵尸，骷髅，苦力怕，骷髅，铁傀儡等命令召唤对应怪物，召唤对应怪物分别需要2，3，2，4，8个鬼火。之前的召唤上限和律令命令依然存在

+++ Assets/Scripts/Manager/FighingManagers/FightingManager.cs
@@ -524,9 +524,9 @@ public class FightingManager : MonoBehaviourPunCallbacks
 
                     if (pveMode)
                     {
-                        AddPlayerDataValue(player.uid, "coin", PVEManager.Instance.difficulty);
+                        AddPlayerDataValue(player.uid, "coin", (int)PVEManager.Instance.difficulty);
                         
-                        emeraldCount = PVEManager.Instance.difficulty;
+                        emeraldCount = (int)PVEManager.Instance.difficulty;
                     }
                 }
             }
